TableTools v0.5: First version of supporting a List<class> like structure
A "table" is technically based on TStringList. It can be created from a csv-string or TStringList and be transformed back into a csv string or TStringList. It has separated "columns" which may be configured to only allow certain data types. Data in the table can be accessed via row index and column index or column name. Columns can be added at a later stage. Data in the table can be sorted via column, reacting differently for different data types. Filters can be set on columns to filter the data. Filters don't destroy data, but when they are removed the respective rows appear again (like in Excel).

overall: standardized working with fieldenclosers and separaters
FormulaParser: used the new TableTools for handling CSVs
XEditTools: some new functions for reading the workshop menu and FLST in general
StringTools: added functionality to ignore empty entries in StringList-Functions, migrated check if something is numeric from formula parser to StringTools